Simplified and extended – ex. better tooltip!
New feature: Now uses the payment link title as the button tooltip if such a title exist.

Simplifications: The click handler has been delegated to the new 0.6.0 version of moz-urlbarbutton which is now given the payment link URL and thus the checkLinks() method has been removed as links are only searched for in a single location now.